+++ lib/logidze/meta.rb
@@ -3,18 +3,19 @@
 module Logidze # :nodoc:
   # Provide methods to attach meta information
   module Meta
-    def with_meta(meta, &block)
-      MetaTransaction.wrap_with(meta, &block)
+    def with_meta(meta, transactional: true, &block)
+      wrapper = transactional ? MetaWithTransaction : MetaWithoutTransaction
+      wrapper.wrap_with(meta, &block)
     end
 
-    def with_responsible(responsible_id, &block)
+    def with_responsible(responsible_id, transactional: true, &block)
       return yield if responsible_id.nil?
 
       meta = {Logidze::History::Version::META_RESPONSIBLE => responsible_id}
-      with_meta(meta, &block)
+      with_meta(meta, transactional: transactional, &block)
     end
 
-    class MetaTransaction # :nodoc:
+    class MetaWrapper # :nodoc:
       def self.wrap_with(meta, &block)
         new(meta, &block).perform
       end
@@ -29,13 +30,11 @@ module Logidze # :nodoc:
       end
 
       def perform
-        return if block.nil?
+        raise ArgumentError, "Block must be given" unless block
         return block.call if meta.nil?
 
-        ActiveRecord::Base.transaction { call_block_in_meta_context }
+        call_block_in_meta_context
       end
-
-      private
 
       def call_block_in_meta_context
         prev_meta = current_meta
@@ -44,10 +43,9 @@ module Logidze # :nodoc:
 
         pg_set_meta_param(current_meta)
         result = block.call
-        pg_reset_meta_param(prev_meta)
-
         result
       ensure
+        pg_reset_meta_param(prev_meta)
         meta_stack.pop
       end
 
@@ -60,20 +58,49 @@ module Logidze # :nodoc:
         Thread.current[:meta]
       end
 
-      def pg_set_meta_param(value)
-        encoded_meta = connection.quote(ActiveSupport::JSON.encode(value))
-        connection.execute("SET LOCAL logidze.meta = #{encoded_meta};")
+      def encode_meta(value)
+        connection.quote(ActiveSupport::JSON.encode(value))
       end
 
       def pg_reset_meta_param(prev_meta)
         if prev_meta.empty?
-          connection.execute("SET LOCAL logidze.meta TO DEFAULT;")
+          pg_clear_meta_param
         else
           pg_set_meta_param(prev_meta)
         end
       end
     end
 
-    private_constant :MetaTransaction
+    class MetaWithTransaction < MetaWrapper # :nodoc:
+      private
+
+      def call_block_in_meta_context
+        connection.transaction { super }
+      end
+
+      def pg_set_meta_param(value)
+        connection.execute("SET LOCAL logidze.meta = #{encode_meta(value)};")
+      end
+
+      def pg_clear_meta_param
+        connection.execute("SET LOCAL logidze.meta TO DEFAULT;")
+      end
+    end
+
+    class MetaWithoutTransaction < MetaWrapper # :nodoc:
+      private
+
+      def pg_set_meta_param(value)
+        connection.execute("SET logidze.meta = #{encode_meta(value)};")
+      end
+
+      def pg_clear_meta_param
+        connection.execute("SET logidze.meta TO DEFAULT;")
+      end
+    end
+
+    private_constant :MetaWrapper
+    private_constant :MetaWithTransaction
+    private_constant :MetaWithoutTransaction
   end
 end